requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-environ==0.10.0
psycopg2-binary==2.9.9
Pillow==10.1.0
drf-spectacular==0.26.5
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.0